ENH: SlurmWatchdog now includes time elapsed before it was started
This should fix #34. In most cases, this won't make a difference,
but if for any reason significant time elapsed before the Watchdog
is initiated, it will now subtract that time from the available walltime.

(Note that we also could rewrite everything to always query the elapsed time,
but that would hammer slurm at every step, which seems bad.)