Remove maximum version for ase
As far as I can tell, a version cap for `ase` was introduced to
avoid some issues with `socketio`. 
They appear to be fixed: https://gitlab.com/ase/ase/-/issues/1145. 

Currently, this version constraint makes it impossible to install `vibes`
with other packages that have a `ase` `master` dependency, which is inconvenient.